Rename form_apply stimulus controller to form_disable_submit and remove target
