+++ app/views/event_groups/manage_entrant_photos.html.erb
@@ -16,14 +16,14 @@
 
   <%= form_with(model: @presenter.event_group,
                 url: update_entrant_photos_event_group_path,
-                html: { method: :patch, data: { controller: "form-apply" } }) do |f| %>
+                html: { method: :patch, data: { controller: "form-disable-submit" } }) do |f| %>
     <div class="mb-3 text-center">
       <div class="dropzone dropzone-default dz-clickable"
            data-controller="dropzone"
            data-dropzone-max-file-size="1"
            data-dropzone-max-files="50"
            data-dropzone-accepted-files="image/*"
-           data-action="dropzone:success->form-apply#enableSubmitButton">
+           data-action="dropzone:success->form-disable-submit#enableSubmitButton">
         <%= f.file_field :entrant_photos, multiple: true, direct_upload: true, data: { "dropzone-target" => "input" } %>
         <div class="dropzone-msg dz-message needsclick text-muted">
           <h3 class="dropzone-msg-title">Drag here to upload or click here to browse</h3>
@@ -33,7 +33,7 @@
     </div>
 
     <div class="mb-3 text-center">
-      <%= f.submit "Upload files", class: "btn btn-primary", data: { turbo: false, "form-apply-target" => "submitButton" } %>
+      <%= f.submit "Upload files", class: "btn btn-primary", data: { turbo: false } %>
       <%= link_to "Back to entrants", entrants_event_group_path(@presenter.event_group), class: "btn btn-outline-secondary" %>
     </div>
   <% end %>
