+++ app/views/events/_course_gpx_form.html.erb
@@ -8,7 +8,7 @@
     <div class="modal-body">
       <%= form_with model: [event.course],
                     url: attach_course_gpx_event_group_event_path(event.event_group, event),
-                    data: { controller: "form-apply", turbo_frame: "_top" } do |f| %>
+                    data: { controller: "form-disable-submit", turbo_frame: "_top" } do |f| %>
         <% if current_user.errors.any? %>
           <div id="error_explanation">
             <ul>
@@ -25,7 +25,7 @@
                data-dropzone-max-file-size="1"
                data-dropzone-max-files="1"
                data-dropzone-accepted-files=".gpx"
-               data-action="dropzone:success->form-apply#enableSubmitButton"
+               data-action="dropzone:success->form-disable-submit#enableSubmitButton"
           >
             <%= f.file_field :gpx, direct_upload: true, data: { "dropzone-target" => "input" } %>
             <div class="dropzone-msg dz-message needsclick text-muted">
@@ -36,7 +36,7 @@
         </div>
 
         <div class="mb-3 text-center">
-          <%= f.submit "Save GPX", class: "btn btn-primary", data: { "form-apply-target" => "submitButton" } %>
+          <%= f.submit "Save GPX", class: "btn btn-primary" %>
         </div>
       <% end %>
     </div>
