+++ app/views/user_settings/password.html.erb
@@ -9,7 +9,7 @@
     <%= render "sidebar" %>
     <div class="col py-3">
       <h4>Change Email</h4>
-      <%= form_with(model: current_user, url: user_settings_update_path, html: { method: :put, data: { controller: "form-apply" } }) do |f| %>
+      <%= form_with(model: current_user, url: user_settings_update_path, html: { method: :put, data: { controller: "form-disable-submit" } }) do |f| %>
         <% if current_user.errors.any? %>
           <div id="error_explanation">
             <ul>
@@ -25,7 +25,7 @@
             <%= f.email_field :email, class: "form-control", placeholder: "Email Address" %>
           </div>
           <div class="col-12 col-md mb-3">
-            <%= f.submit "Save Changes", class: "btn btn-primary", data: { turbo: false, "form-apply-target" => "submitButton" } %>
+            <%= f.submit "Save Changes", class: "btn btn-primary", data: { turbo: false } %>
           </div>
         </div>
       <% end %>
@@ -34,7 +34,7 @@
       <hr/>
 
       <h4>Change Password</h4>
-      <%= form_for(current_user, url: registration_path(:user), html: { method: :put }, data: { controller: "form-apply" }) do |f| %>
+      <%= form_for(current_user, url: registration_path(:user), html: { method: :put }, data: { controller: "form-disable-submit" }) do |f| %>
         <div class="mb-3">
           <%= f.label :old_password %>
           <%= f.password_field :current_password, autocomplete: "off", class: "form-control", placeholder: "Current Password" %>
@@ -52,7 +52,7 @@
         </div>
 
         <div class="mb-3">
-          <%= f.submit "Save Changes", class: "btn btn-primary", data: { turbo: false, "form-apply-target" => "submitButton" } %>
+          <%= f.submit "Save Changes", class: "btn btn-primary", data: { turbo: false } %>
         </div>
       <% end %>
